refactor: enhance according to change request of reveiw
