+++ src/components/__test__/Subscribe.test.js
@@ -1,6 +1,6 @@
 import renderer from "react-test-renderer";
 
-import Subscribe from "./../Subscribe";
+import Subscribe from "../Subscribe";
 
 it("renders correctly", () => {
     const tree = renderer.create(<Subscribe />).toJSON();
