+++ src/pages/index.jsx
@@ -5,7 +5,6 @@ import * as React from "react";
 
 import Layout from "@/layout/Layout";
 
-import Subscribe from "../components/Subscribe";
 export default function HomePage() {
     const { t } = useTranslation("common");
 
@@ -20,7 +19,6 @@ export default function HomePage() {
                     العربية
                 </Link>
             </div>
-            <Subscribe />
         </Layout>
     );
 }
